requirements.txt
numpy
pandas
scikit-learn
fastapi
uvicorn
opentelemetry-exporter-gcp-trace
python-json-logger
